requirements.txt
Django>=3.0,<4.0;
cx_Oracle;
djangorestframework;
flask;
psycopg2-binary;
django-cors-headers;